+++ app/models/invoice.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Text, DateTime, Boolean
+from sqlalchemy import Column, Integer, Text, DateTime, Boolean, BigInteger
 from app.database import Base
 
 class Invoice(Base):
@@ -6,7 +6,7 @@ class Invoice(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     replacement_id = Column(Integer, nullable=True, index=True)
-    order_id = Column(Integer, nullable=True)
+    order_id = Column(BigInteger, nullable=True)
     companyVatCode = Column(Text, nullable=True)
     seriesName = Column(Text, nullable=True)
     client = Column(Text, nullable=True)
