+++ app/models/reverse_invoice.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Text
+from sqlalchemy import Column, Integer, Text, BigInteger
 from app.database import Base
 
 class Reverse_Invoice(Base):
@@ -6,12 +6,10 @@ class Reverse_Invoice(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     replacement_id = Column(Integer, nullable=True, index=True)
-    order_id = Column(Integer, nullable=True, index=True)
+    order_id = Column(BigInteger, nullable=True, index=True)
     companyVatCode = Column(Text, nullable=True)
     seriesName = Column(Text, nullable=True)
     factura_number = Column(Text, nullable=True, index=True)
     storno_number = Column(Text, nullable=True, index=True)
     post = Column(Integer, nullable=True)
     user_id = Column(Integer, index=True, nullable=True)
-
-
